+++ src/test/java/TestUtils.java
@@ -5,7 +5,7 @@ import java.util.HashMap;
 import java.util.Map;
 
 public class TestUtils {
-    public static Pair<Map<Long, Long>, Map<Long, Long>> generateAB(long base, long n, long absT) {
+    public static Pair<Map<Long, Long>, Map<Long, Long>> generateAB(int base, int n, long absT) {
         Map<Long, Long> A = new HashMap<>();
         Map<Long, Long> B = new HashMap<>();
         long x = 1;
